# Backend/CIT-Forums/src/main/resources/static/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CIT Forums API",
    "description": "API documentation for CIT Forums platform",
    "version": "1.0.0",
    "contact": {
      "name": "CIT Forums Admin",
      "email": "admin@citforums.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/admin/monitor/stats": {
      "get": {
        "summary": "Get detailed system statistics",
        "description": "Provides comprehensive statistics about the system including user counts, content counts, and server metrics",
        "tags": [
          "System Monitor"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "System statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/monitor/memory": {
      "get": {
        "summary": "Get memory usage statistics",
        "description": "Provides detailed memory usage statistics including heap and non-heap memory",
        "tags": [
          "System Monitor"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Memory usage statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryUsageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "SystemStatusResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "UP"
          },
          "serverTime": {
            "type": "string",
            "format": "date-time"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "uptime": {
            "type": "string"
          },
          "forumCount": {
            "type": "integer",
            "format": "int64"
          },
          "threadCount": {
            "type": "integer",
            "format": "int64"
          },
          "commentCount": {
            "type": "integer",
            "format": "int64"
          },
          "userCount": {
            "type": "integer",
            "format": "int64"
          },
          "systemMetrics": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "MemoryUsageResponse": {
        "type": "object",
        "properties": {
          "heapMemoryUsage": {
            "type": "object",
            "properties": {
              "init": {
                "type": "integer",
                "format": "int64",
                "description": "Initial heap memory allocation in bytes"
              },
              "used": {
                "type": "integer",
                "format": "int64",
                "description": "Currently used heap memory in bytes"
              },
              "committed": {
                "type": "integer",
                "format": "int64",
                "description": "Committed heap memory in bytes"
              },
              "max": {
                "type": "integer",
                "format": "int64",
                "description": "Maximum heap memory in bytes"
              },
              "usagePercentage": {
                "type": "number",
                "format": "double",
                "description": "Percentage of heap memory used"
              }
            }
          },
          "nonHeapMemoryUsage": {
            "type": "object",
            "properties": {
              "init": {
                "type": "integer",
                "format": "int64"
              },
              "used": {
                "type": "integer",
                "format": "int64"
              },
              "committed": {
                "type": "integer",
                "format": "int64"
              },
              "max": {
                "type": "integer",
                "format": "int64"
              }
            }
          },
          "totalMemory": {
            "type": "integer",
            "format": "int64",
            "description": "Total memory available to JVM in bytes"
          },
          "freeMemory": {
            "type": "integer",
            "format": "int64",
            "description": "Free memory available to JVM in bytes"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
} 